Trimestre 4/Instructora_Isaura/evidencia de aprendizaje/Actividad 14/requirements.txt
Flask==2.0.1
Flask-Bcrypt==0.7.1
PyMySQL==1.0.2
Werkzeug==2.0.3
python-dotenv==0.19.2